+++ partiql-eval/src/eval/evaluable.rs
@@ -1,4 +1,5 @@
 use crate::env::basic::MapBindings;
+use crate::error::EvaluationError;
 use crate::eval::expr::EvalExpr;
 use crate::eval::{EvalContext, EvalPlan};
 use itertools::Itertools;
@@ -10,6 +11,21 @@ use std::cmp::{max, min, Ordering};
 use std::collections::{HashMap, HashSet};
 use std::fmt::Debug;
 use std::rc::Rc;
+
+#[macro_export]
+macro_rules! take_input {
+    ($expr:expr, $ctx:expr) => {
+        match $expr {
+            None => {
+                $ctx.add_error(EvaluationError::IllegalState(
+                    "Error in retrieving input value".to_string(),
+                ));
+                return None;
+            }
+            Some(val) => val,
+        }
+    };
+}
 
 /// `Evaluable` represents each evaluation operator in the evaluation plan as an evaluable entity.
 pub trait Evaluable: Debug {
@@ -163,7 +179,12 @@ impl Evaluable for EvalJoin {
         let lhs_values = self.left.evaluate(ctx);
         let left_bindings = match lhs_values {
             Some(Value::Bag(t)) => *t,
-            _ => panic!("Left side of FROM source should result in a bag of bindings"),
+            _ => {
+                ctx.add_error(EvaluationError::IllegalState(
+                    "Left side of FROM source should result in a bag of bindings".to_string(),
+                ));
+                return None;
+            }
         };
 
         // Current implementations follow pseudocode defined in section 5.6 of spec
@@ -272,7 +293,10 @@ impl Evaluable for EvalJoin {
                 });
             }
             EvalJoinKind::Full | EvalJoinKind::Right => {
-                todo!("Full and Right Joins are not yet implemented for `partiql-lang-rust`")
+                ctx.add_error(EvaluationError::NotYetImplemented(
+                    "FULL and RIGHT JOIN".to_string(),
+                ));
+                return None;
             }
         };
         Some(Value::Bag(Box::new(output_bag)))
@@ -306,7 +330,7 @@ pub trait AggregateFunction {
     /// Provides the next value for the given `group`.
     fn next_value(&mut self, input_value: &Value, group: &Tuple);
     /// Returns the result of the aggregation function for a given `group`.
-    fn compute(&self, group: &Tuple) -> Value;
+    fn compute(&self, group: &Tuple) -> Result<Value, EvaluationError>;
 }
 
 #[derive(Debug)]
@@ -330,7 +354,7 @@ impl AggregateFunction for AggFunc {
         }
     }
 
-    fn compute(&self, group: &Tuple) -> Value {
+    fn compute(&self, group: &Tuple) -> Result<Value, EvaluationError> {
         match self {
             AggFunc::Avg(v) => v.compute(group),
             AggFunc::Count(v) => v.compute(group),
@@ -435,10 +459,13 @@ impl AggregateFunction for Avg {
         }
     }
 
-    fn compute(&self, group: &Tuple) -> Value {
-        match self.avgs.get(group).expect("Expect group to exist in avgs") {
-            (0, _) => Null,
-            (c, s) => s / &Value::Decimal(rust_decimal::Decimal::from(*c)),
+    fn compute(&self, group: &Tuple) -> Result<Value, EvaluationError> {
+        match self.avgs.get(group) {
+            None => Err(EvaluationError::IllegalState(
+                "Expect group to exist in avgs".to_string(),
+            )),
+            Some((0, _)) => Ok(Null),
+            Some((c, s)) => Ok(s / &Value::Decimal(rust_decimal::Decimal::from(*c))),
         }
     }
 }
@@ -482,12 +509,13 @@ impl AggregateFunction for Count {
         }
     }
 
-    fn compute(&self, group: &Tuple) -> Value {
-        Value::from(
-            self.counts
-                .get(group)
-                .expect("Expect group to exist in counts"),
-        )
+    fn compute(&self, group: &Tuple) -> Result<Value, EvaluationError> {
+        match self.counts.get(group) {
+            None => Err(EvaluationError::IllegalState(
+                "Expect group to exist in counts".to_string(),
+            )),
+            Some(val) => Ok(Value::from(val)),
+        }
     }
 }
 
@@ -530,11 +558,13 @@ impl AggregateFunction for Max {
         }
     }
 
-    fn compute(&self, group: &Tuple) -> Value {
-        self.maxes
-            .get(group)
-            .expect("Expect group to exist in sums")
-            .clone()
+    fn compute(&self, group: &Tuple) -> Result<Value, EvaluationError> {
+        match self.maxes.get(group) {
+            None => Err(EvaluationError::IllegalState(
+                "Expect group to exist in maxes".to_string(),
+            )),
+            Some(val) => Ok(val.clone()),
+        }
     }
 }
 
@@ -577,11 +607,13 @@ impl AggregateFunction for Min {
         }
     }
 
-    fn compute(&self, group: &Tuple) -> Value {
-        self.mins
-            .get(group)
-            .expect("Expect group to exist in mins")
-            .clone()
+    fn compute(&self, group: &Tuple) -> Result<Value, EvaluationError> {
+        match self.mins.get(group) {
+            None => Err(EvaluationError::IllegalState(
+                "Expect group to exist in mins".to_string(),
+            )),
+            Some(val) => Ok(val.clone()),
+        }
     }
 }
 
@@ -624,11 +656,13 @@ impl AggregateFunction for Sum {
         }
     }
 
-    fn compute(&self, group: &Tuple) -> Value {
-        self.sums
-            .get(group)
-            .expect("Expect group to exist in sums")
-            .clone()
+    fn compute(&self, group: &Tuple) -> Result<Value, EvaluationError> {
+        match self.sums.get(group) {
+            None => Err(EvaluationError::IllegalState(
+                "Expect group to exist in sums".to_string(),
+            )),
+            Some(val) => Ok(val.clone()),
+        }
     }
 }
 
@@ -670,10 +704,15 @@ impl EvalGroupBy {
 impl Evaluable for EvalGroupBy {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
         let group_as_alias = &self.group_as_alias;
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         match self.strategy {
-            EvalGroupingStrategy::GroupPartial => todo!(),
+            EvalGroupingStrategy::GroupPartial => {
+                ctx.add_error(EvaluationError::NotYetImplemented(
+                    "GROUP PARTIAL".to_string(),
+                ));
+                None
+            }
             EvalGroupingStrategy::GroupFull => {
                 let mut groups: HashMap<Tuple, Vec<Value>> = HashMap::new();
                 for v in input_value.into_iter() {
@@ -698,8 +737,15 @@ impl Evaluable for EvalGroupBy {
                         // tuple
                         let mut agg_results: Vec<(&str, Value)> = vec![];
                         for aggregate_expr in &self.aggregate_exprs {
-                            let agg_result = aggregate_expr.func.compute(&k);
-                            agg_results.push((aggregate_expr.name.as_str(), agg_result));
+                            match aggregate_expr.func.compute(&k) {
+                                Ok(agg_result) => {
+                                    agg_results.push((aggregate_expr.name.as_str(), agg_result))
+                                }
+                                Err(err) => {
+                                    ctx.add_error(err);
+                                    return Null;
+                                }
+                            }
                         }
                         agg_results
                             .into_iter()
@@ -747,7 +793,7 @@ impl EvalPivot {
 
 impl Evaluable for EvalPivot {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         let tuple: Tuple = input_value
             .into_iter()
@@ -861,7 +907,7 @@ impl EvalFilter {
 
 impl Evaluable for EvalFilter {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         let filtered = input_value
             .into_iter()
@@ -906,7 +952,7 @@ impl EvalHaving {
 
 impl Evaluable for EvalHaving {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         let filtered = input_value
             .into_iter()
@@ -986,7 +1032,7 @@ impl EvalOrderBy {
 
 impl Evaluable for EvalOrderBy {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         let mut values = input_value.into_iter().collect_vec();
         values.sort_by(|l, r| self.compare(l, r, ctx));
@@ -1008,7 +1054,7 @@ pub(crate) struct EvalLimitOffset {
 
 impl Evaluable for EvalLimitOffset {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         let empty_bindings = Tuple::new();
 
@@ -1077,7 +1123,7 @@ impl EvalSelectValue {
 
 impl Evaluable for EvalSelectValue {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         let ordered = input_value.is_ordered();
 
@@ -1115,7 +1161,7 @@ impl EvalSelect {
 
 impl Evaluable for EvalSelect {
     fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+        let input_value = take_input!(self.input.take(), ctx);
 
         let ordered = input_value.is_ordered();
 
@@ -1158,8 +1204,8 @@ impl EvalSelectAll {
 }
 
 impl Evaluable for EvalSelectAll {
-    fn evaluate(&mut self, _ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+    fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
+        let input_value = take_input!(self.input.take(), ctx);
 
         let ordered = input_value.is_ordered();
 
@@ -1221,8 +1267,8 @@ impl EvalDistinct {
 }
 
 impl Evaluable for EvalDistinct {
-    fn evaluate(&mut self, _ctx: &dyn EvalContext) -> Option<Value> {
-        let input_value = self.input.take().expect("Error in retrieving input value");
+    fn evaluate(&mut self, ctx: &dyn EvalContext) -> Option<Value> {
+        let input_value = take_input!(self.input.take(), ctx);
         let ordered = input_value.is_ordered();
 
         let values = input_value.into_iter().unique();