+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 ### Added
+- Ability to add and view errors during evaluation with partiql-eval's `EvalContext`
 ### Fixes
 
 ## [0.4.1] - 2023-05-25